+++ tests/Helper/OutputHelperTest.php
@@ -29,12 +29,15 @@ class OutputHelperTest extends TestCase
 
     public function setUp(): void
     {
-        file_put_contents(static::$ou, '');
+        file_put_contents(static::$ou, '', LOCK_EX);
     }
 
     public static function tearDownAfterClass(): void
     {
-        unlink(static::$ou);
+        // Make sure we clean up after ourselves:
+        if (file_exists(static::$ou)) {
+            unlink(static::$ou);
+        }
     }
 
     public function test_show_arguments()
